# TimeSheet/src/main/resources/OpenApi.yml
openapi: 3.0.0
info:
  title: Sample API
  description: Optional multiline or single-line description in [CommonMark](http://commonmark.org/help/) or HTML.
  version: 0.1.9
...
paths:
...
  /client:
    get:
      summary: This gets all teammembers
      description: This method return list of TeamMembers
      tags:
        - "Client"
      operationId: getClients
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientDTO'
        "400":
          description: BAD REQUEST
    post:
      summary: This method creates Client
      tags:
        - "Client"
      operationId: insertClient
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientDTO'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDTO'
        "400":
          description: BAD REQUEST
    put:
      summary: This method update exist Client
      tags:
        - "Client"
      operationId: updateClient
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientDTO'
      responses:
        "200":
          description: SUCCESS
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDTO'
        "400":
          description: BAD REQUEST
        "404":
          description: NOT FOUND
  /client/{clientid}:
    get:
      parameters:
        - in: path
          name: clientid
          required: true
          schema:
            type: integer
      summary: Returns Client if exist
      tags:
        - "Client"
      operationId: getClient
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDTO'
        "400":
          description: BAD REQUEST
        "404":
          description: NOT FOUND
    delete:
      parameters:
        - in: path
          name: clientid
          required: true
          schema:
            type: integer
      summary: Delete client if exist
      tags:
        - "Client"
      operationId: deleteClient
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDTO'
        "400":
          description: BAD REQUEST
        "404":
          description: NOT FOUND
  /country:
    get:
      summary: This gets all countries
      description: This method return list of countries
      tags:
        - "Country"
      operationId: getCountries
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CountryDTO'
        "400":
          description: BAD REQUEST
    post:
      summary: This method creates Country
      tags:
        - "Country"
      operationId: insertCountry
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CountryDTO'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryDTO'
        "400":
          description: BAD REQUEST
    put:
      summary: This method update exist Country
      tags:
        - "Country"
      operationId: updateCountry
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CountryDTO'
      responses:
        "200":
          description: SUCCESS
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryDTO'
        "400":
          description: BAD REQUEST
        "404":
          description: NOT FOUND
  /country/{countryid}:
    get:
      parameters:
        - in: path
          name: countryid
          required: true
          schema:
            type: integer
      summary: Returns Country if exist
      tags:
        - "Country"
      operationId: getCountry
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryDTO'
        "400":
          description: BAD REQUEST
        "404":
          description: NOT FOUND
    delete:
      parameters:
        - in: path
          name: countryid
          required: true
          schema:
            type: integer
      summary: Delete country if exist
      tags:
        - "Country"
      operationId: deleteCountry
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryDTO'
        "400":
          description: BAD REQUEST
        "404":
          description: NOT FOUND
...
components:
  schemas:
...
    ClientDTO:
      description: This is the Client object.
      type: object
      properties:
        id:
          description: Id of client
          type: integer
        name:
          description: Name of client
          type: string
          pattern: ^(?!\s*$).+
        address:
          description: Address of client
          type: string
        city:
          description: City where client lives
          type: string
        zipCode:
          description: Zip code of city where client lives
          type: string
        countryid:
          description: Country id
          type: integer
      required:
        - name
    CountryDTO:
      description: This object represents parameters for country
      type: object
      properties:
        countryid:
          description: id of country
          type: integer
        name:
          description: name of country
          type: string
          pattern: ^(?!\s*$).+
      required:
        - name
...